// prisma/generated/type-graphql/models/Brands.d.ts
export declare class Brands {
    id: bigint;
    created_at?: Date | null;
    code?: string | null;
    name?: string | null;
    country?: string | null;
    state?: string | null;
    isPublic?: string | null;
    city?: string | null;
    product?: bigint | null;
    description?: string | null;
    media?: string | null;
    mediamanager?: string | null;
    products?: string | null;
    workspaces?: string | null;
}